requirements.txt
Flask
PyMuPDF
requests
python-dotenv
openpyxl
beautifulsoup4